changes to manage better the gui

+++ calendar_app_clean/lib/c-frontend/b-group-section/screens/show-groups/group_profile/profile_alert_dialog_actions.dart
@@ -93,8 +93,22 @@ List<Widget> buildProfileDialogActions(
           );
 
           if (confirm) {
-            await groupManagement.groupService.deleteGroup(group.id);
-            if (context.mounted) Navigator.pop(context);
+            if (group.ownerId == user.id) {
+              try {
+                await groupManagement.groupService.deleteGroup(group.id);
+                if (context.mounted) Navigator.pop(context);
+              } catch (e) {
+                ScaffoldMessenger.of(context).showSnackBar(
+                  SnackBar(content: Text("Error: ${e.toString()}")),
+                );
+              }
+            } else {
+              ScaffoldMessenger.of(context).showSnackBar(
+                const SnackBar(
+                    content:
+                        Text("Only the group owner can delete this group.")),
+              );
+            }
           }
         },
         child: Text(
